Commonize InvalidationTracker
Move the bulk of the TRIGGER logic in Room's InvalidationTracker to an internal class in the common source set called TriggerBasedInvalidationTracker. The public APIs of InvalidationTracker will then delegate to the internal implementation allowing functionality to be shared across platforms.

Asynchronous related logic has been migrated to Coroutines using Room's new CoroutineScope and provided CoroutineContext.

Two new APIs are added to the common version of the tracker (and available to Android too), 'subscribe' and 'unsubscribe' suspend functions to add and remove observers from the tracker.

Bug: 299168035
Bug: 309990302
Bug: 309996304
Test: Existing++
Relnote: "Add suspend APIs for subscribing and unsubscribing InvalidationTracker.Observer with the RoomDatabase tracker."
Change-Id: If2a1f5b2442c6772b67afcae92996d1a8a9ff2ac