Update AbstractDraggableNode definition.
Moving AbstractDraggableNode to mean a Modifier.Node that implements drag gesture recognition. This means that this node will only care about gesture recognition, therefore we're passing to the node only the inputs necessary for gesture recognition. Reverse direction is an implementation detail of Draggables (Draggable2D, Draggable1D and AnchoredDraggable) and is not relevant for gesture recognition so it was abstracted away from the AbstractDraggableNode. In this change we also update the name to DragGestureNode to make more clear the definition of a "a node that detects drag gestures".

Test: N/A
Change-Id: I9e494528e1713f6ce16a6f0c1710968aa3046ba0