Correct nullability of Continuation type bound
The `kotlin.coroutines.Continuation` parameter is added to the Java
implementation of `suspend` functions automatically by Kotlin. Its
type argument is supposed to be the return type of the original Kotlin
suspend function. Due to a bug in Metalava the nullability of that type
was ignored and it was always treated as non-null. That bug has now
been fixed and this change corrects the AndroidX API signature files.

Relnote: N/A
Bug: 325454211
Test: ./gradlew
Change-Id: Iad3146c4b11d422dcfa492a1ef234f12d33cc7fd