Correct nullability of vararg in an inline function
Kotlin vararg parameters are always non-null, i.e. the array which is
passed to the vararg parameter is never null. Unfortunately, due to a
bug in Metalava (and possibly one in the underlying psi library too)
varargs in inline functions were treated as nullable. The bug in
Metalava has been fixed and this is updating the AndroidX API signature
files to match.

Relnote: N/A
Bug: 325454211
Test: ./gradlew
Change-Id: Ic56e14f893819368113d05b7e713da429753734b