Correct nullability of Continuation and bound in suspend lambda
The `kotlin.coroutines.Continuation` parameter is added to the Java
implementation of `suspend` lambdas automatically by Kotlin. Its
type argument is supposed to be the return type of the original Kotlin
suspend function. Due to a bug in Metalava the nullability of that type
and of the `Continuation` type, were ignored and they were always
treated as non-null and nullable respectively, when it should have been
the other way around. That bug has now been fixed and this change
corrects the AndroidX API signature files.

Relnote: N/A
Bug: 325454211
Test: ./gradlew
Change-Id: I37b8e771401c2ef84aae9e4790da759c69859b1a