Skip temporary variable desugarded from destructuring expression
For example,

  val (first, second, third) = (list as List<Int>)

is desugarded to

  val varHash = list as List<Int>
  val first = varHash.component1()
  val second = varHash.component2()
  val thrid = varHash.component3()

Here, varHash, the temporary variable, has primitive in collection.
K1 UAST missed this due to type resolution error, whereas
K2 UAST has a proper type on that variable, resulting in an alarm.

Since the counterpart in androidx, e.g., IntList, does not have
componentN(), we cannot suggest that as a quickfix. Instead, it seems
better to skip such temporary variables in the first place.

Bug: 328122546
Test: unit test added, and tested w/ and w/o K2 UAST
Change-Id: Ia79796892cd62cfd89a11ad0f0734cab8970d514