Exchanging GSON for Kotlin Serialization JSON in room-migration
Refactor room-migration to use kotlinx-serialization-json instead of the GSON library. Both libraries have the necessary features for room-migration but Kotlin's is KMP compatible.

The main difference in the refactor is how polymorphic type EntityBundle is handled, specifically in kotlinx-serialization the properties of the parent class must be abstract or otherwise it is considered duplicate properties, this lead to the inheritance of FtsEntityBundle and EntityBundle to be refactored into implementations of a sealed class named BaseEntityBundle. To help with polymorphic types kotlinx-serialization uses a type classifier, but we configure encoding to not emit the classifier as that encodes internal class details and instead, similar to GSON we use a content inspector to decided which type to decode, specifically if the 'ftsVersion' key is present we decode the JSON object into a FtsEntityBundle otherwise it is decoded into an EntityBundle.

Kotlin serialization also supports optional (or missing) key-value pairs, this is important as there are certain entries that are not present in older generates schemas, such as IndexBundle's orders. These are handled by using a default value in the constructor parameter. Meanwhile 'transient' properties (those that are not encoded into the JSON) are also supported either via an annotation, or by having no backing field (such as delegate properties).

Lastly, room-migration is solely used by room-compiler and room-testing but due to historical reasons its APIs where marked as @RestrictTo with LIBRARY_GROUP_PREFIX but are in fact restricted to LIBRARY_GROUP only. No APIs in this project are ever used in generated code or other androidx libraries. This CL fixes this by removing updating the @RestrictTo which allows further modifications to the APIs to conform to Kotlin's serialization requirements and to cleanup GSON used constructors.

Note that the migration tests in Room's integration project contain an extensive set of schema files that have been generated by old versions of the compiler, the fact that these schemas are decoded and the tests pass are a testament of the correctness of this CL, thus the minimal change to tests. Similarly that no current schema files changed is a signal that schema equality if working as expected.

Bug: 299168035
Bug: 317908685
Test: Existing
Change-Id: I43a7f9b9672efc656628b52fe2012d2ef75e1a8f