webkit: Update MimeUtil javascript type
IETF deprecated application/javascript in favor of text/javascript so
updating to reflect that.

The primary reason we added this was to support the asset loader library
in AndroidX for WebView, and WebView supports this type so this is a
safe update.

The testGuessMimeType test started failing because Android now has the
javascript mime - and because the test checks the OS first, this causes
us to fail. In practice, this means we are mainly updating the test to
pass, and then updating the mime type for Android versions that do not
support the Javascript mime type.

One issue is that Android did previously report application/javascript
so we need to support both cases in our test. Updating the test to
just check for both possible values since they are both valid anyway.

Test: gradlew :webkit:integration-tests:instrumentation:connectedAndroidTest
Bug: 329107332
Change-Id: Ib2ce640f9aa08602e5b845e8a47cd43aba3435e1