Move snapshot inside of lazy measure policy
Makes sure the measure result is applied in a global snapshot.

Previously, whole of lazy layout measure block was wrapped with a snapshot for consistency. This resulted in Snapshot.apply() and remeasure request while the lazy layout node was requested (as the snapshot was applied during measure). The lazy layout, however, updates scrollPosition and other internal states from `applyMeasureResult`, the last step of lazy measure logic, and expects some of those changes to invalidate measure for lazy node again. As this started happening during the measurement of lazy layout node, the measure request was ignored.
By moving snapshot apply before the `applyMeasureResult` call, the remeasure request happens after the node was measured, and is no longer ignored.

Fixes: 320936743
Test: manually with the repro in the bug, as scroll/recomposition sequence is not reproducible in test conditions

Change-Id: Ie09aae6feefdfa79425ee2707fd3d67faa02c00b