Fix the issue in beauty mode in extensions testlib that could fail the test
The beauty mode in extensions testlib returns supported resolutions that
incudes both JPEG and YUV size. but it implements the CaptureProcessor
so that YUV is used to be the capture format and yet it uses the
maximum JPEG as the size to configure the YUV image reader.

To fix the issue, we only return supported YUV size to avoid the issue

Bug: 318451053
Test: extensions/ImageCaptureTest
Change-Id: Iea508779148f4f6ee3614cb5305b0f1f4cd5320b