Fix incorrect view reuse test case
The existing test case wasn't testing the expected case. Instead of the
movable content under test moving, Compose would detect one of its
siblings moving instead. Because of how reuse groups have changed in
the latest compiler + runtime combo, this test would start to see
the movable content itself being moved because of how the `if`
statement in the column's repeat block is compiled into a replace group.

This CL updates the test to ensure it tests the same recomposition
pathway before and after the compiler change. The general case involves
detaching and re-attaching the View, which feels like a realistic
behavior to expect. We could theoretically avoid removing and
re-attaching the LayoutNode if it's going to end up in the same parent,
but that isn't a guarantee of MovableContent so we shouldn't assert it.

Test: AndroidViewTest.testViewIsReused_whenMoved
Change-Id: I05d78c284147d01bf393c0e267d5ced51eeaf991