Begin migrating docs samples resolution to use dependencies, not hardcode
The immediate advantage of this is addressing b/329424152. We can now ensure
that tasks declare a dependency on their samples.

Removes checking samples from CheckTipOfTreeDocsTask.

For docs-public, this only works with libraries that have been published since the sample changes landed.
For docs-tip-of-tree, this works with all libraries, and all samples() have been removed from its build.gradle.

Test: passes in presubmit; with one more removal from docs-public/build.gradle it would fail. Without the library build.gradle fixes, docs-tip-of-tree now fails.
Change-Id: Iad938e98a8741d6aab8adf7b9558a7554ef15453