Hide PaneScaffoldDirective constructor
We expect the fields of PaneScaffoldDirective will keep expanding.
Introduces a copy pattern and suggest client to use it instead of using
the constructor directly so we can have more flexibility on ensuring
the API compatibility. It also makes it easier for dev who only want to
customize a few fields of it.

Relnote: hide PaneScaffoldDirective constructor
Test: unit test
Change-Id: I9291f4b5f096e9f6266ed13081a6c9dab2427259