Introducing Delegating Adapters
Type of uiAdapter that enables refreshing sessions with a
SandboxedSdkView without the View owner's involvement. The uiAdapter can switch views
between other SandboxedUiAdapters.

To refresh a session with a SandboxedSdkView the adapter triggers an
event for a secondary client, and fully establishes a session with the
secondaryclient of the view before switching it to primary.

Bug: b/340471011
Relnote: "Introduce Delegating Adapters which are used by UiProviders
    for switching to a new delegate uiProvider"
Test: manually tested refresh ads on the test app

Change-Id: I5f1c501d1f4018d0f5eba2bd569eafe065e824aa