Add default value to expectAssertionError
The test-utils function expectError has a parameter `expectError:
Boolean` with a default value of true, to accommodate parameterized
tests that expect an error in some cases and not in other.

`expectAssertionError` is a convenience function for when the error
class you expect is an AssertionError, complete with an `expectError:
Boolean` parameter. However, this parameter doesn't have a default
value where it makes much more sense to have one.

Bug: N/A
Test: presubmit
Change-Id: I828e9f377e07697eaa52118649310430a5e1a24a