Fix DragAndDropTargetNode Delegation Lifecycle
This changes the behavior to delegate and undelegate according to the lifecycle. Updates to the target are now destructive to the underlying DragAndDropModifierNode, recreating the node if needed.
This is important when  `DragAndDropTargetNode` instances are being re-used, e.g. in lazy layouts.

Fixes: 325512895
Relnote: Fixed a bug where `Modifier.dragAndDropTarget()` could reference stale data in certain scenarios with Modifier.Node re-use.
Test: DragDropTargetTest
Change-Id: I05bb1a7dea19acf4234f88fe988f51e3c2822f4b