Fix child transition added after end
Fixes: 330717902

An animation that was added to a completed child transition
had its start time set to 0, even if the parent transition
was still running. When the child transition started again,
that animation would be instantly completed because its
start time was set improperly.

This CL uses the root transition's play time rather than the
child transitions play time.

Test: manual, new test

Change-Id: I292237400ee75294342cd5175091c08bc5a30c7c