CreateEntry gets BiometricPromptData
This adds the BiometricPromptData for the Create Flow, and allows
providers to showcase the biometric screen directly with credential
manager if they opt in. Therefore, this reduces the flow to a single tap
should they choose to not follow up with any other steps when having
users create a passkey.

Bug: 325469395
Test: JUnit Tests immediately chained
Relnote: "Providers will be able to utilize a flow that embeds the
biometric screen in the credential manager selector on creating a
credential."

Change-Id: I89ecd1d6396b2fc7d1a2b110dcc7c0349d28539b