Migrating ArrayQueryResultAdapter to use driver APIs.
This CL refactors the ArrayQueryResultAdapter to delegate
to the ListQueryResultAdapter, since SQLStatement API's
do not allow access to the row count as Android's Cursor
did. This workaround is also beneficial in reducing
boilerplate code as well as removing the special handling
of the nullable array types, since the List adapter can
handle this scenario.

The only downside is the conversion from list to an array
is not ideal, but we can circle back to it if we have a
better solution in the future.

Bug: 299168035
Bug: 333903173
Test: DaoKotlinCodeGenTest & BaseSimpleQueryTest
Change-Id: I4fe55163496005ed33de32b61e7c4d39c7a647df