Make findNode for KClass reified
NavGraph #get and #contains delegates to the #findNode KClass version. Due to changes to findNode, get and contain are refactored to reified as well to pass the Type Parameter. However #get and #contains, being operator functions, requires at least one parameter. So the KClass for these two options are still expected to be passed in as `MyClass::class` parameter.

Test: ./gradlew navigation:navigation-common:cC
Bug: 188693139
Change-Id: I8c4d2d17bb12bf71b3059650b1f163df192e0528