Include self for FocusOwner Key and Rotary Traversal
In the Modifier.Node world, a node could be both a focus target as well as a key input/soft keyboard interception/rotary input node. When traversing the ancestors, we want to include the active focus target node in the traversal as it could be the node we are looking for.

This will let us flatten hierarchies and delegate to focus target from e.g. Clickable instead, which would be both a focus target and key input node.

Test: FocusTargetAttachDetachTest
Change-Id: Ica5c964d244daadf321e13224b9b22c817cfb10e