Implement NavController getBackStackEntry for Safe Args
Added getBackStackEntry overloads to retrieve entry with a serializable KClass or with a serializable Object.

For KClass version, there are no arguments values to consider so we can get entry directly based on its id. For Object version, we need to generate the route with args filled and then get entry with that.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 188693139
Relnote: "Added NavController getBackStackEntry APIs for retrieving entries with a serializable KClass or with a serializable Object. The target NavDestination to be retrieved must have been created with route from KClass."
Change-Id: Icb70f85a4fdfd7b8c472bd4a282e446e86ae8dde