Make AnchoredDraggable's targetValue threshold-less
We are preparing to move configuration like the thresholds out of AnchoredDraggableState. targetValue was previously considering positional thresholds in the calculation, but this can be moved to the component level.

Test: anchoredDraggable_targetState_updatedWithDeltaDispatch
Relnote: AnchoredDraggableState's targetValue now does *not* consider positional thresholds anymore. It now has an implicit threshold of 50%, meaning that the targetValue changes at the midpoint between two anchors.
Change-Id: I82c2c160f078ebdee15c74e0b0016ed60446cc94