Refactor test app to use a DrawerLayout to separate flows
Introduces a base fragment that contains helper methods for
easier testing, as well as abstract methods to make the drawer
opening/closing UX cleaner. This change is essentially a no-op,
but allows for easier separation of manual testing flows in the
future. New flows can be tested by subclassing BaseFragment,
adding a new item to action_menu.xml, and starting the new
fragment when the item is clicked in MainActivity.

Test: Manual
Bug: 331564598
Change-Id: Id3b1da4e0151f187b4596c58132275ec4855ff7b