Add pausable composition
A PausableComposition is a sub-composition that allows spreading
the initial composition multiple calls. This, for example, can
be used by LazyColumn's precomposition to spread the composition
of an item that is likely to be scrolled on to screen across
multiple frames preventing large items from causing the current
from starting.

Pausing is currently not enabled as it requires compiler plugin
changes that are currently in development.

Test: New test and existing tests: ./gradlew :compose:r:r:tDUT
Fixes: 328817808
Relnote: """Added PausableComposition which allows creating
a subcomposition that can be paused during composition and
and applied asynchronously to the composition. Compiler support
is required for pausing which is currently in development.
"""

Change-Id: I3394b7ebd90932ca451c7d55df7f286329de7656