Gracefully handle remote process death for UI sessions
Since the UI provider for remote sessions lives in a separate process,
it is possible for the remote process to die while the host process
remains alive. This causes issues when the client library tries to
call IBinders owned by the remote process.

To mitigate this, we can do two things. We will call
onRemoteSessionError when the remote process dies. We also wrap all
calls to the remote session to catch any RemoteExceptions caused by
dead objects.

This is tested by unloading all SDKs while a remote view is shown. The
StateChangeListener will remove the SandboxedSdkView from the view tree, but this will not cause any errors when the remote adapter tries to call close() on the remote session controller.

Test: Manual
Bug: 331776180
Change-Id: I3bc1d6cdda8b8b5d247bc61b63bbcd31b16d0b83