Fix SQL statement identification for Android drivers
Android's implementation has some limitations on how queries are executed based on returning rows or not. Initially android.database.DatabaseUtils.getSqlStatementType() seemed enough but it lacks identification of common table expression (CTEs). Since we are only interested if a statement returns a row or not, then this change makes it so we use our own simple prefix check.

Bug: 300660920
Test: BaseConformanceTest
Change-Id: I37954945a96da1f13cbef309aae092843e544b60