Fixed the extensions/ImageCaptureTest failure on Pixel 2 Emulator and re-enable the tests
root cause: when using glTexImage2D to copy the buffer of Image to
OpenGL, eglSwapBuffers is blocked and never returns.

solution: copy the buffer of the Image and also use the correct
texture id for glBindTexture

Also re-enable the tests in ImageCaptureTest/AdvancedSessionProcessorTest
and BasicExtenderSessionProcessorTest as they are all caused by this
issue.

Bug: 331617278
Bug: 331871255
Test: extensions/ImageCaptureTest
Change-Id: I3d32f99f92ab348ed081205286843c899ddf2f2f