Update navDeepLink dsl to take basePath
This ensures basePath is set for safe args deeplinks. Any uriPattern that is set within the builder lambda will override the generated uriPattern.

The builder requires `noinline` or else IDE throws error `Illegal usage of inline-parameter 'deepLinkBuilder...`

Note on lint: navDeepLink must not have empty builder lambdas in the past. Now it is permitted if there is TypeParameter in the case of safe args where the base path for uriPattern is passed in as a param rather than set as a uriPattern within builder.

Test: ./gradlew navigation:navigation-common:cC
Bug: 188693139
Bug: 334995120
Relnote: "Safe Args navDeepLink dsl now takes in basePath to ensure base path is set for the final generated uriPattern. Any uriPattern that is set within the builder lambda will override the generated uriPattern."
Change-Id: I8d3948948e68b65f05c5bb943daba3deda64db2e