Synchronize Torch test verification to prevent race conditions
The current torch test result verification process has
a race condition. Verification updates and the arrival
of new capture results can overlap, causing some results
to be missed and leading to test failures.

Introduce a synchronized block to ensure that the verification
process and the list that stores new results are treated
atomically. This will guarantee that all incoming capture
results are consistently verified.

Bug: 333741831
Test: Camera2CapturePipelineTest
Change-Id: I18dafd01d1cbafe991513afda02ee90ef6661c5f