Apply API feedback on LayoutCoordinates & LookaheadScope
Changed `isPositionedByParentWithDirectManipulation` to `introducesFrameOfReference`. However, in this case, for the default case it's expected to be true. And, when placed under `withDirectManipulation{}` it's expected to be false.

This also splits the added `excludeDirecctManipulationOffset` flag into a separate method: `positionInLocalFrameOfReference` for when the parameter would otherwise be true.

Also changed mentions of `withDirectManipulation` to `withCurrentFrameOfReference` (and corresponding derivatives).

Relnote: "Renamed `isPositionedByParentWithDirectManipulation` to `introducesFrameOfReference`. Note that it now has the reverse effect, meaning, that by default, most LayoutCoordinates introduce a frame of reference, and, only when placed under direct manipulation the property will be false. To query position with only those that introduce a frame of reference, use `positionInLocalFrameOfReference(...)`. Or `positionInLocalLookaheadFrameOfReference` from a LookaheadScope."

Bug: 330540544
Test: n/a, passing
Change-Id: Ifc5f78c683543035e13ff727edf14a79075b1a84