Avoid crashing when perfetto trace processor is slow to start
In b/329145808 it seems that perfetto can sometimes be slow to start
and, while the process is running, query an end point might return an
error, in which case a `FileNotFoundException` is thrown. Since we
already check for the process to be running, it's ok to "catch" this
state of the perfetto trace processor and wait for it for the timeout
duration.

Bug: 329145808
Test: ./gradlew benchmark:benchmark-macro:test
Relnote: Avoid crashing when perfetto trace processor is slow to start
Change-Id: I988418daa69ec33343364f7ad7c5fca26a39a5fa