Generate Java lambda syntax when jvmTarget > 8
This CL introduces the utility function InvokeWithLambdaParameter() that generate a code block to invoke a function whose last parameter is a functional type. The utility will correctly generate the right code in Java (with an anonymous class if targeting older than Java 8 or with lambda syntax if targeting Java 8 or greater) and Kotlin.

Bug: 322387497
Test: DaoWriterTest
Change-Id: Ib9f5ea310fc550a04262b35d2ff1103420e05cb8