Fix performance regression in new layers implementation
In complex apps we can have a lot of graphics layers being active at the same time. In some of the benchmarks, for example we switch between complex screens and we have 200 layers from previous screen still active, then we create 200 more layers for the next screen and only then we release the old 200 layers. The previous implementation in LayerManager was always iterating over all the active layers and drawing them every time we add new layer. It is extremely inefficient on this scale of having a lot of layers at once.
What we can do instead is to not do the persistence straight away but instead wait for the frame to fully settle and only then we draw all the active layers only once for the. We can achieve it by posting a message into the handler.
I verified that for many benchmarks this change fixes a regression introduced by merging aosp/3003872 where we started using new layers everywhere.

Test: run benchmarks and AndroidGraphicsLayerTest on different api levels
Change-Id: I5dc5579fec0dfd360a70a66d90ed99e2f9dbf2b1