Reverse AnchoredDraggable deltas by default in RTL layouts
Adds a Modifier.anchoredDraggable overload without a reverseDirection parameter that looks at LocalLayoutDirection and the orientation and reverses the deltas by default for LayoutDirection.RTL && Orientation.Horizontal.
This is the behavior we want by default in e.g. Switch and SwipeToDismissBox.

Relnote: Modifier.anchoredDraggable now reverses drag deltas by default for components with a Horizontal orientation when used in an RTL layout.
Test: anchoredDraggable_reverseDirection_true_reversesDeltas, anchoredDraggable_reverseDirection_defaultValue_reversesDeltasInRTL
Bug: 332368717
Change-Id: I3c6d950cedc30b1790e09e140d3d83d19a5d312b