Fixes indication inside clickable not being properly disposed inside movableContentOf / ReusableContent
When reused / moved, we now clean up and dispose any previously created indication node / lazily created InteractionSource. This means that when attached again, the node will be clean of any stale state. Previously we would reuse the node / interaction source, which meant that if reused / moved while focused for example, the focus highlight could still be visible in some cases, and would be stuck.

Also adds test cases for hoverable for the same scenarios (the equivalent tests already exist for focusable)

Fixes: b/338127256
Test: ClickableTest
Test: HoverableTest
Change-Id: Iceff46b9a4c8f67d8288e876b478b3f2a34098ca