fix(non linear font scaling): fix inconsistencies when the font scale is less than the lowest defined table
Some OEMs and form factors use scales that are lower than 1.15, the
lowest defined table. e.g. Wear at 1.06. This causes an issue at some
font sizes where the font will be bigger at linear 1.06 vs the next
nonlinear scale option, 1.15. This looks weird when you are dragging
the font scale slider to increase the scale but the font gets smaller.

We fix it by interpolating between the linear 1x scale and the lowest
defined table, so that the font size is monotonically increasing as
scales increase.

Fix: 336720383
Fix: 337049808
Test: ./gradlew :compose:ui:ui-unit:connectedDebugAndroidTest

Change-Id: Icbae34e19e771ec2f55e74badcca3b4c90d55d25