Improve Pager default snapping animation.
Specifying a visibility threshold for Pager improves the alignment of the visual representation of the end of a snapping operation with the actual PagerState interpretation of a scrolling Pager. It takes now twice less time between the time the Pager "looks settled" with the indication that it stopped scrolling "isScrollInProgress=false" which should signal to the user that they can interact with the screen.

Test: Previous tests should pass.
Fixes: 338710348
Bug: 303807950
Relnote: N/A
Change-Id: I9627a48d53604a0919951fa4dbb74717ba30bd28