Undo changes made to type variable bounds
We changed the nullability of type variable bounds to Nullable if it's Platform in KSP in aosp/3008521. After aosp/3083585, this is no longer necessary as we now generate nullable type names from most platform type usages. This CL is to revert the change but keep the tests.

Note that we still add an Any? (with Nullable nullability instead of Platform as we can't create that) to the super types of type variables if we don't find it as KSP2 only adds Any! to the bounds if there is no bounds specified. This is to make the superTypes behavior more consistent with KAPT.

Also note that @NotNull bounds in Java are still nullable in KSP2. This is safer so we're keeping this behavior.

Test: XExecutableTypeTest
Change-Id: I9837ad766d8183dd184c3b0788b0e92537d01114