Links Api: revert adding TextLinkStyles to TextStyle
This CL addresses the API feedback but unlike aosp/3079405 doesn't move the TextLinkStyles to the TextStyle.
API changes since latest update:
- Remove the TextLinkStyles from the TextStyle
- Add back the styling argument to the AnnotatedString.Companion.fromHtml
- Remove merge methods from the TextLinkStyles and LinkAnnotation (can be done using all public ctors)

So material won't do any styling to the links at the moment. If

Fixes: 335640940
Test: TextTest, BasicTextLinkTest

Relnote: "Removed TextLinkStyles from the TextStyle class. Instead, TextLinkStyles is part of the LinkAnntation constructor and the AnnotatedString.fromHtml method"

Change-Id: I90b2b73e126d9c1106c223de823dda8babaf6708