Update docs of MainTestClock and ComposeUiTest
The documentation of MainTestClock was still claiming that on Android
measure and layout passes are solely driven by the Choreographer, but we
are performing traversals already since r.android.com/2260389. This
should be very confusing to developers.

Traces of these claims were also found in ComposeUiTest and
ComposeTestRule. Furthermore, it seems that the documentation of the
methods in ComposeTestRule haven't been updated to match the same
methods in ComposeUiTest, even though they use the same implementation.

Bug: 340911033
Test: N/A
Change-Id: I0edcbfb7878e10c1c8d0128018e3c0c012f3ec6d