Clear pending render counts in clearParamQueues
Failing to reset the count of pending commits leads to redundant render calls when the surface is recreated. This results in extra calls to onDrawMultiBufferedLayer with an empty params list (more than the one expected), or in additional front renders that skip the user onDrawFrontBufferedLayer callback (because that's called within a block passed to ParamQueue.next that does not get executed unless there's actually an item in the queue).

The same issue happens on clear, which uses the same underlying method to clear pending draws before triggering a render of the multi-buffered layer.

This fix might help with a rendering stall where (apparently) the count of pending commits can be spuriously positive while no commits are actually pending. One way for this to happen is for commit to be called while mMultiBufferedRenderer is null, that increments mCommitCount, but no render is actually kicked off. mMultiBufferedRender is null while the app is paused (assuming the default SurfaceView lifecycle). clearParamQueues is called when the surface is updated on resume, but that doesn't currently clear mCommitCount.

Getting tests that reliably failed before the fix and pass elsewhere required fixing another bug where deferred render requests did not get processed on release when the cancelPending parameter is set to false. It needs to push those up-front so the sequence point set by GLThread.stop(false) is after those callbacks.

Fix: 342154090
Fix: 342160701
Bug: 344867857
Test: Unit tests that fail before the code change and pass after
Change-Id: I42da1458f6c4dbbbe45a988b4f8a9eb278bf32ad