Propagate errors correctly and retry on camera-specific errors
We currently only restart `Camera2CameraController` when the last error
is `ERROR_CAMERA_DEVICE`. This is unnecessarily stringent. When a camera
becomes available (`onCameraAvailable`), it means it is available to be
used and we should be able to retry unless we have a graph configuration
error, which would suggest issues with the `CameraGraph.Config` and that
a retry would not succeed.

Another quirk is added such that when `CameraAvailable` is received on a
started `Camera2CameraController`, attmept a restart, too. It seems that
noncompliant camera HALs might not fire `onDisconnected` or `onError`
callbacks when the camera is disconnected.

Bug: 340583180
Test: FocusMeteringDeviceTest
Change-Id: I2afeb0673592fbc0f06030e68ad2a6ab451a65dd