Partially reenable the ignored transformation tests
Reenable the part of test which are not affected. Added TODO with the bug filed to roboletric.

After the upgrade of roboletric from 4.12.1 to 4.12.2, the return value of view#getPivotX or view#getPivotY when its corresponding setter is never called, their
is  no longer the central location of the view but zero.

Bug: 342225240

Change-Id: Ic3e70334f1840d45fd7bf1864bfc1e7aca7162ae